requirements.txt
# Core dependencies
mysql-connector-python>=8.2.0
requests>=2.31.0
python-dateutil>=2.8.2
Flask>=3.0.0

# Testing
pylint>=3.3.8
pylint-google-style-guide-imports-enforcing>=1.3.1
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-mock>=3.11.1
pytest-benchmark>=4.0.0
coverage[toml]>=7.3.0

# Code quality
black>=23.7.0
isort>=5.12.0
pydocstyle>=6.3.0

# Security
semgrep>=1.132.0
bandit>=1.7.5
safety>=2.3.0